requirements.txt
Flask_SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
Flask==2.0.0
Flask-Sockets
Werkzeug==2.0.0
requests
flask_cors
Pillow >= 8.3.2
gunicorn
numpy==1.22
beautifulsoup4
plotly
pandas
networkx
matplotlib
scipy
scikit-learn
pymysql
shapely
cartoGRAPHs